+++ rtp_llm/models_py/triton_kernels/common/activation.py
@@ -30,7 +30,7 @@ def _silu_and_mul_kernel(
     # Meta-parameter for tuning
     BLOCK_SIZE_N: tl.constexpr,
 ):
-    pid_b = tl.program_id(axis=0)  # Batch dimension
+    pid_b = tl.program_id(axis=0).to(tl.int64)  # Widen before row * stride.
     pid_n_block = tl.program_id(axis=1)  # N-dimension block
 
     input_row_start_ptr = input_ptr + pid_b * input_row_stride
